fix: making firestore emulator work for common unit tests
