+++ microservices/sample_service/src/routes/user_test.py
@@ -7,7 +7,8 @@ import datetime
 
 # disabling pylint rules that conflict with pytest fixtures
 # pylint: disable=unused-argument,redefined-outer-name,unused-import
-from common.testing.firestore_emulator import client_with_emulator, firestore_emulator, clean_firestore
+from common.testing.firestore_emulator import firestore_emulator, clean_firestore
+from common.testing.client_with_emulator import client_with_emulator
 
 from common.models import User
 import mock
